Update usage of default PAs
The PA 'default' implementation is just providing symbols
on the assumption that it is not provided by an actual PA implementation.

Currently it is declared and used at the PA level, which means that it is
up to the PA to use or not the default PA.
This patch update daemons to have the declaration at their level, so that
the default PA is used anyway.
It also prevents dependency on the AT PA for the localhost target regarding
AirVantage, Secure Storage and fwupdate.

Change-Id: I0859f3ad0b4a8fa648f3ac67e1730a17a02e3aee
Resolves: trac:4654