Provide way to update simu values through ConfigTree
This allows to update values returned by the simu PA
through the ConfigTree.

For instance, to set the PSN, you can do:
 config set /simulation/modem/info/psn TESTPSN

It also makes it possible to handle new values entries
in pa_*_simu.c files and registering this to simuConfig.
The sole use so far is in pa_info_simu.c .

Resolves: LE-6531
Change-Id: Ic23ab3bf9d788bb9a9f7e1f120d7a4ff25286fae